Major structural cleanup finished.
Remove as much unused code as possible, conform to MVC, incorporate
express router, dotenv.
