+++ .gitignore
@@ -8,3 +8,4 @@ bashScript5.sh
 bashScript6.sh
 bashScript7.sh
 *.csv
+.env
