+++ db/SQL/Queries/couchbaseQueries.sql
