+++ db/seed/inheritedNoSqlGen.js
