+++ package.json
@@ -67,6 +67,7 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "csv-write-stream": "^2.0.0",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "fs": "0.0.1-security",
     "lorem-ipsum": "^2.0.3",
