--- server/environmentalVariables.js
@@ -1,17 +1,0 @@
-// Service and Development mode environmental variables
-const environmentalVariables = {
-  IP_ADDRESS: '127.0.0.1',
-  IP_ADDRESS_E: '127.0.0.1',
-  IP_ADDRESS_K: '127.0.0.1',
-  DATABASE_LOCAL_ADDRESS: 'localhost'
-};
-
-// Deployment mode environmental variables
-// const environmentalVariables = {
-//   IP_ADDRESS: '54.183.137.155',
-//   IP_ADDRESS_E: '52.14.208.55',
-//   IP_ADDRESS_K: '18.224.229.28',
-//   DATABASE_LOCAL_ADDRESS: 'mongo',
-// };
-
-module.exports = environmentalVariables;
